requirements.txt
altgraph==0.17
asgiref==3.4.1
astroid==2.5.3
colorama==0.4.4
Django==3.2.6
djangorestframework==3.13.1
future==0.18.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pefile==2021.5.24
Pillow==8.3.1
psycopg2==2.9.2
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.7.4
pytz==2021.1
pywin32-ctypes==0.2.0
sqlparse==0.4.1
style==1.1.0
toml==0.10.2
update==0.0.1
whitenoise==5.3.0
wrapt==1.12.1
